Add torchscript export unittest
